requirements.txt
asgiref==3.7.2
black==24.3.0
click==8.1.7
colorama==0.4.6
Django==5.0.3
django-filter==24.2
djangorestframework==3.15.1
flake8==7.0.0
Markdown==3.6
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pycodestyle==2.11.1
pyflakes==3.2.0
sqlparse==0.4.4
tzdata==2024.1
